+++ src/Request/Filters/Decoders/FilterDecoderInterface.php
@@ -2,10 +2,10 @@
 
 namespace Somnambulist\Bundles\ApiBundle\Request\Filters\Decoders;
 
+use Somnambulist\Bundles\ApiBundle\Request\Contracts\Searchable;
 use Somnambulist\Bundles\ApiBundle\Request\Filters\Expression\CompositeExpression;
-use Somnambulist\Bundles\ApiBundle\Request\FormRequest;
 
 interface FilterDecoderInterface
 {
-    public function decode(FormRequest $request): CompositeExpression;
+    public function decode(Searchable $request): CompositeExpression;
 }