+++ src/Request/Filters/Decoders/NestedArrayFilterDecoder.php
@@ -2,12 +2,12 @@
 
 namespace Somnambulist\Bundles\ApiBundle\Request\Filters\Decoders;
 
+use Somnambulist\Bundles\ApiBundle\Request\Contracts\Searchable;
 use Somnambulist\Bundles\ApiBundle\Request\Filters\Decoders\Behaviours\ConvertOperator;
 use Somnambulist\Bundles\ApiBundle\Request\Filters\Decoders\Behaviours\ConvertStringToArray;
 use Somnambulist\Bundles\ApiBundle\Request\Filters\Expression\CompositeExpression;
 use Somnambulist\Bundles\ApiBundle\Request\Filters\Expression\Expression;
 use Somnambulist\Bundles\ApiBundle\Request\Filters\Expression\ExpressionInterface;
-use Somnambulist\Bundles\ApiBundle\Request\FormRequest;
 
 use function array_key_exists;
 
@@ -21,7 +21,7 @@ class NestedArrayFilterDecoder implements FilterDecoderInterface
     use ConvertStringToArray;
     use ConvertOperator;
 
-    public function decode(FormRequest $request): CompositeExpression
+    public function decode(Searchable $request): CompositeExpression
     {
         $filters     = $request->filters();
         $expressions = $this->getCompositeContainer($filters);